Refactor applicant switch to sole notifications
