+++ src/test/java/uk/gov/hmcts/divorce/citizen/event/CitizenSwitchedToSoleTest.java
@@ -10,7 +10,8 @@ import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.Event;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.ccd.sdk.type.AddressGlobalUK;
-import uk.gov.hmcts.divorce.citizen.notification.SwitchToSoleNotification;
+import uk.gov.hmcts.divorce.citizen.notification.Applicant1SwitchToSoleNotification;
+import uk.gov.hmcts.divorce.citizen.notification.Applicant2SwitchToSoleNotification;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseInvite;
@@ -18,6 +19,7 @@ import uk.gov.hmcts.divorce.divorcecase.model.HelpWithFees;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
 import uk.gov.hmcts.divorce.idam.IdamService;
+import uk.gov.hmcts.divorce.notification.NotificationDispatcher;
 import uk.gov.hmcts.divorce.solicitor.service.CcdAccessService;
 import uk.gov.hmcts.reform.idam.client.models.User;
 import uk.gov.hmcts.reform.idam.client.models.UserDetails;
@@ -29,10 +31,10 @@ import javax.servlet.http.HttpServletRequest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.NO;
 import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.YES;
 import static uk.gov.hmcts.divorce.citizen.event.CitizenSwitchedToSole.SWITCH_TO_SOLE;
 import static uk.gov.hmcts.divorce.divorcecase.model.ApplicationType.SOLE_APPLICATION;
@@ -50,7 +52,7 @@ import static uk.gov.hmcts.divorce.testutil.TestDataHelper.validJointApplicant1C
 class CitizenSwitchedToSoleTest {
 
     @Mock
-    private SwitchToSoleNotification switchToSoleNotification;
+    private Applicant1SwitchToSoleNotification applicant1SwitchToSoleNotification;
 
     @Mock
     private CcdAccessService ccdAccessService;
@@ -60,6 +62,12 @@ class CitizenSwitchedToSoleTest {
 
     @Mock
     private IdamService idamService;
+
+    @Mock
+    private Applicant2SwitchToSoleNotification applicant2SwitchToSoleNotification;
+
+    @Mock
+    private NotificationDispatcher notificationDispatcher;
 
     @InjectMocks
     private CitizenSwitchedToSole citizenSwitchedToSole;
@@ -98,9 +106,8 @@ class CitizenSwitchedToSoleTest {
 
         final AboutToStartOrSubmitResponse<CaseData, State> response = citizenSwitchedToSole.aboutToSubmit(caseDetails, caseDetails);
 
-        verify(switchToSoleNotification).sendApplicant1SwitchToSoleNotificationToApplicant1(caseData, caseDetails.getId());
-        verify(switchToSoleNotification).sendApplicant1SwitchToSoleNotificationToApplicant2(caseData, caseDetails.getId());
-        verifyNoMoreInteractions(switchToSoleNotification);
+        verify(notificationDispatcher).send(applicant1SwitchToSoleNotification, caseData, caseDetails.getId());
+        verifyNoMoreInteractions(notificationDispatcher);
 
         assertThat(response.getData().getApplicationType()).isEqualTo(SOLE_APPLICATION);
     }
@@ -130,9 +137,8 @@ class CitizenSwitchedToSoleTest {
 
         final AboutToStartOrSubmitResponse<CaseData, State> response = citizenSwitchedToSole.aboutToSubmit(caseDetails, caseDetails);
 
-        verify(switchToSoleNotification).sendApplicant1SwitchToSoleNotificationToApplicant1(caseData, caseDetails.getId());
-        verify(switchToSoleNotification).sendApplicant1SwitchToSoleNotificationToApplicant2(caseData, caseDetails.getId());
-        verifyNoMoreInteractions(switchToSoleNotification);
+        verify(notificationDispatcher).send(applicant1SwitchToSoleNotification, caseData, caseDetails.getId());
+        verifyNoMoreInteractions(notificationDispatcher);
 
         assertThat(response.getData().getApplicationType()).isEqualTo(SOLE_APPLICATION);
         assertThat(response.getData().getCaseInvite().getAccessCode()).isNull();
@@ -161,40 +167,9 @@ class CitizenSwitchedToSoleTest {
 
         final AboutToStartOrSubmitResponse<CaseData, State> response = citizenSwitchedToSole.aboutToSubmit(caseDetails, caseDetails);
 
-        verify(switchToSoleNotification).sendApplicant2SwitchToSoleNotificationToApplicant1(caseData, caseDetails.getId());
-        verify(switchToSoleNotification).sendApplicant2SwitchToSoleNotificationToApplicant2(caseData, caseDetails.getId());
-        verifyNoMoreInteractions(switchToSoleNotification);
-
-        assertThat(response.getData().getApplicationType()).isEqualTo(SOLE_APPLICATION);
-    }
-
-    @Test
-    void givenApplicant2ScreenHasMarriageBrokenIsNoThenOnlyApplicant1NotificationIsSentForApplicant1SwitchToSole() {
-        final long caseId = 1L;
-        final CaseDetails<CaseData, State> caseDetails = new CaseDetails<>();
-        CaseData caseData = validJointApplicant1CaseData();
-        caseData.getApplication().setApplicant2ScreenHasMarriageBroken(NO);
-        setValidCaseInviteData(caseData);
-
-        caseDetails.setData(caseData);
-        caseDetails.setId(caseId);
-
-        when(httpServletRequest.getHeader(AUTHORIZATION))
-            .thenReturn("token");
-
-        final var userDetails = UserDetails.builder()
-            .email(TEST_USER_EMAIL)
-            .id("app1")
-            .build();
-
-        when(idamService.retrieveUser(anyString()))
-            .thenReturn(new User("token", userDetails));
-
-
-        final AboutToStartOrSubmitResponse<CaseData, State> response = citizenSwitchedToSole.aboutToSubmit(caseDetails, caseDetails);
-
-        verify(switchToSoleNotification).sendApplicant1SwitchToSoleNotificationToApplicant1(caseData, caseDetails.getId());
-        verifyNoMoreInteractions(switchToSoleNotification);
+        verify(notificationDispatcher).send(applicant2SwitchToSoleNotification, caseData, caseDetails.getId());
+        verifyNoMoreInteractions(notificationDispatcher);
+        verifyNoInteractions(applicant1SwitchToSoleNotification);
 
         assertThat(response.getData().getApplicationType()).isEqualTo(SOLE_APPLICATION);
     }
